requirements.txt
appdirs
beautifulsoup4
buildozer
colorama
cx-freeze
cython
docutils
huey
jinja2
pep517
pip-chill
pygments
pyperclip
pyqt5-stubs
pyqtwebengine
pyserial
pytest-cov
pytest-mock
pytest-qt
qtbot
requests
# huey>=2.4
# PyQt5~=5.15.5
# pyserial==3.4
# requests~=2.26.0
# pyperclip~=1.8.2
# cx_freeze
# beautifulsoup4~=4.11.1
# pytest~=5.2.2
